+++ tests/crowdsourcing/tasks/fast_acute/test_no_self_chat_fast_acute.py
@@ -20,7 +20,9 @@ try:
 
     from parlai.crowdsourcing.tasks.fast_acute.run_no_self_chat import (
         NoSelfChatFastAcuteExecutor,
-        BLUEPRINT_TYPE,
+    )
+    from parlai.crowdsourcing.tasks.fast_acute.fast_acute_blueprint import (
+        FAST_ACUTE_NO_SELF_CHAT_BLUEPRINT_TYPE,
     )
     from parlai.crowdsourcing.tasks.fast_acute.util import AbstractFastAcuteTest
 
@@ -68,7 +70,7 @@ try:
             ]
             # TODO: clean this up when Hydra has support for recursive defaults
             self._set_up_config(
-                blueprint_type=BLUEPRINT_TYPE,
+                blueprint_type=FAST_ACUTE_NO_SELF_CHAT_BLUEPRINT_TYPE,
                 task_directory=self.ACUTE_EVAL_TASK_DIRECTORY,
                 overrides=self._get_common_overrides(root_dir) + no_self_chat_overrides,
             )
